use the new prompt format
